Source: Refactor some stuff
bench: 6697391
