+++ Source/history.h
@@ -12,5 +12,6 @@ void update_capture_history_moves(thread_t *thread,
 void update_continuation_history_moves(thread_t *thread, searchstack_t *ss,
                                        moves *quiet_moves, int best_move,
                                        uint8_t depth);
+int16_t get_conthist_score(thread_t *thread, searchstack_t *ss, int move);
 
 #endif
