+++ Source/see.h
@@ -1,0 +1,8 @@
+#ifndef SEE_H
+#define SEE_H
+
+#include "structs.h"
+
+int SEE(position_t *pos, int move, int threshold);
+
+#endif
